Use set for etcd register and supported unregister.
