--- config.go
@@ -1,7 +1,0 @@
-package main
-
-type etcdConf struct {
-	SelfIp   string
-	SelfPort int
-	EtcdAddr string
-}
